refactor(slack): share thread block text extraction
