+++ extensions/slack/src/monitor/message-handler/prepare-content.ts
@@ -3,12 +3,10 @@ import type { WebClient as SlackWebClient } from "@slack/web-api";
 import { runTasksWithConcurrency } from "openclaw/plugin-sdk/concurrency-runtime";
 import { createLazyRuntimeModule } from "openclaw/plugin-sdk/lazy-runtime";
 import { logVerbose } from "openclaw/plugin-sdk/runtime-env";
-import {
-  normalizeOptionalString,
-  readStringValue as readString,
-} from "openclaw/plugin-sdk/string-coerce-runtime";
+import { normalizeOptionalString } from "openclaw/plugin-sdk/string-coerce-runtime";
 import { formatSlackFileReference } from "../../file-reference.js";
 import type { SlackFile, SlackMessageEvent } from "../../types.js";
+import { chooseSlackPrimaryText, resolveSlackBlocksText } from "../block-text.js";
 import { MAX_SLACK_MEDIA_FILES, type SlackMediaResult } from "../media-types.js";
 import type { SlackThreadStarter } from "../thread.js";
 
@@ -20,36 +18,6 @@ type SlackResolvedMessageContent = {
 const SLACK_MENTION_RESOLUTION_CONCURRENCY = 4;
 const SLACK_MENTION_RESOLUTION_MAX_LOOKUPS_PER_MESSAGE = 20;
 const SLACK_USER_MENTION_RE = /<@([A-Z0-9]+)(?:\|[^>]+)?>/gi;
-
-type SlackTextObject = {
-  text?: unknown;
-};
-
-type SlackRichTextElement = {
-  type?: unknown;
-  text?: unknown;
-  url?: unknown;
-  user_id?: unknown;
-  channel_id?: unknown;
-  usergroup_id?: unknown;
-  name?: unknown;
-  range?: unknown;
-  elements?: unknown;
-};
-
-type SlackBlockLike = {
-  type?: unknown;
-  text?: unknown;
-  elements?: unknown;
-  fields?: unknown;
-  alt_text?: unknown;
-  title?: unknown;
-};
-
-type SlackBlocksText = {
-  text: string;
-  hasRichText: boolean;
-};
 
 const loadSlackMediaModule = createLazyRuntimeModule(() => import("../media.js"));
 
@@ -85,176 +53,6 @@ function renderSlackUserMentions(
     const rendered = renderedMentions.get(userId);
     return rendered ?? full;
   });
-}
-
-function readTextObject(value: unknown): string | undefined {
-  if (!value || typeof value !== "object") {
-    return undefined;
-  }
-  return normalizeOptionalString(readString((value as SlackTextObject).text));
-}
-
-function renderSlackRichTextLeaf(element: SlackRichTextElement): string {
-  switch (element.type) {
-    case "text":
-      return readString(element.text) ?? "";
-    case "link":
-      return readString(element.text) ?? readString(element.url) ?? "";
-    case "user": {
-      const userId = readString(element.user_id);
-      return userId ? `<@${userId}>` : "";
-    }
-    case "channel": {
-      const channelId = readString(element.channel_id);
-      return channelId ? `<#${channelId}>` : "";
-    }
-    case "usergroup": {
-      const usergroupId = readString(element.usergroup_id);
-      return usergroupId ? `<!subteam^${usergroupId}>` : "";
-    }
-    case "broadcast": {
-      const range = readString(element.range);
-      return range ? `<!${range}>` : "";
-    }
-    case "emoji": {
-      const name = readString(element.name);
-      return name ? `:${name}:` : "";
-    }
-    default:
-      return "";
-  }
-}
-
-function renderSlackRichTextElements(elements: unknown): string {
-  if (!Array.isArray(elements)) {
-    return "";
-  }
-  const parts: string[] = [];
-  for (const rawElement of elements) {
-    if (!rawElement || typeof rawElement !== "object") {
-      continue;
-    }
-    const element = rawElement as SlackRichTextElement;
-    switch (element.type) {
-      case "rich_text_section":
-      case "rich_text_preformatted":
-      case "rich_text_quote": {
-        parts.push(renderSlackRichTextElements(element.elements));
-        break;
-      }
-      case "rich_text_list": {
-        const listParts: string[] = [];
-        if (Array.isArray(element.elements)) {
-          for (const child of element.elements) {
-            if (!child || typeof child !== "object") {
-              continue;
-            }
-            const rendered = renderSlackRichTextElements((child as SlackRichTextElement).elements);
-            if (rendered) {
-              listParts.push(rendered);
-            }
-          }
-        }
-        const listText = listParts.join("\n");
-        parts.push(listText);
-        break;
-      }
-      default:
-        parts.push(renderSlackRichTextLeaf(element));
-        break;
-    }
-  }
-  return parts.join("");
-}
-
-function readSlackBlockText(block: unknown): string | undefined {
-  if (!block || typeof block !== "object") {
-    return undefined;
-  }
-  const blockLike = block as SlackBlockLike;
-  switch (blockLike.type) {
-    case "rich_text":
-      return normalizeOptionalString(renderSlackRichTextElements(blockLike.elements));
-    case "section": {
-      const text = readTextObject(blockLike.text);
-      if (text) {
-        return text;
-      }
-      if (Array.isArray(blockLike.fields)) {
-        const fields: string[] = [];
-        for (const field of blockLike.fields) {
-          const fieldText = readTextObject(field);
-          if (fieldText) {
-            fields.push(fieldText);
-          }
-        }
-        return fields.length > 0 ? fields.join("\n") : undefined;
-      }
-      return undefined;
-    }
-    case "header":
-      return readTextObject(blockLike.text);
-    case "context": {
-      if (!Array.isArray(blockLike.elements)) {
-        return undefined;
-      }
-      const parts: string[] = [];
-      for (const element of blockLike.elements) {
-        const text = readTextObject(element);
-        if (text) {
-          parts.push(text);
-        }
-      }
-      return parts.length > 0 ? parts.join(" ") : undefined;
-    }
-    case "image":
-      return (
-        normalizeOptionalString(readString(blockLike.alt_text)) ?? readTextObject(blockLike.title)
-      );
-    case "video":
-      return (
-        readTextObject(blockLike.title) ?? normalizeOptionalString(readString(blockLike.alt_text))
-      );
-    default:
-      return undefined;
-  }
-}
-
-function resolveSlackBlocksText(blocks: unknown[] | undefined): SlackBlocksText | undefined {
-  if (!blocks?.length) {
-    return undefined;
-  }
-  const parts: string[] = [];
-  let hasRichText = false;
-  for (const block of blocks) {
-    if (block && typeof block === "object" && (block as SlackBlockLike).type === "rich_text") {
-      hasRichText = true;
-    }
-    const text = readSlackBlockText(block);
-    if (text) {
-      parts.push(text);
-    }
-  }
-  return parts.length > 0 ? { text: parts.join("\n"), hasRichText } : undefined;
-}
-
-function chooseSlackPrimaryText(params: {
-  messageText: string | undefined;
-  blocksText: SlackBlocksText | undefined;
-}): string | undefined {
-  const { messageText, blocksText } = params;
-  if (!blocksText) {
-    return messageText;
-  }
-  if (!messageText) {
-    return blocksText.text;
-  }
-  if (blocksText.hasRichText && blocksText.text.length > messageText.length) {
-    return blocksText.text;
-  }
-  return blocksText.text.length > messageText.length && blocksText.text.startsWith(messageText)
-    ? blocksText.text
-    : messageText;
 }
 
 function filterInheritedParentFiles(params: {
